requirements.txt
tensorflow>=2.13.0
keras>=2.13.1
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
Pillow>=10.0.0
opencv-python>=4.8.0
tqdm>=4.65.0
plotly>=5.15.0
streamlit>=1.25.0
efficientnet>=1.1.1
pyyaml>=6.0
jupyter>=1.0.0
ipykernel>=6.25.0 